Factor out `run_daemon` and move to library
Factor out
```
pub fn run_daemon(config: ConfigRoot) -> Result<Daemon, Error>
```

from

```
pub fn run(args: &Args) -> Result<(), Error>
```

and move it to the library such that we can call it from tests.
