+++ src/bin/oura/main.rs
@@ -2,21 +2,21 @@ use clap::Parser;
 use std::process;
 
 mod console;
-mod daemon;
+mod run_daemon;
 
 #[derive(Parser)]
 #[clap(name = "Oura")]
 #[clap(bin_name = "oura")]
 #[clap(author, version, about, long_about = None)]
 enum Oura {
-    Daemon(daemon::Args),
+    Daemon(run_daemon::Args),
 }
 
 fn main() {
     let args = Oura::parse();
 
     let result = match args {
-        Oura::Daemon(x) => daemon::run(&x),
+        Oura::Daemon(x) => run_daemon::run(&x),
     };
 
     if let Err(err) = &result {
