+++ src/lib.rs
@@ -3,3 +3,4 @@ pub mod filters;
 pub mod framework;
 pub mod sinks;
 pub mod sources;
+pub mod daemon;
